Feedback

+++ filament/backend/src/vulkan/VulkanDescriptorSetLayoutCache.h
@@ -47,7 +47,7 @@ public:
     // This method is meant to be used with external samplers
     VkDescriptorSetLayout getVkLayout(VulkanDescriptorSetLayout::Bitmask const& bitmasks,
             fvkutils::SamplerBitmask externalSamplers,
-            utils::FixedCapacityVector<std::pair<uint32_t, VkSampler>> immutableSamplers = {});
+            utils::FixedCapacityVector<std::pair<uint64_t, VkSampler>> immutableSamplers = {});
 
 private:
     VkDevice mDevice;
